Custom django auth ldap
Add update_username_or_create_user function.
This requirement comes from my one real project.
"username" field for django auth_user table is primary key, which means it's
required when creating a user. This field is also used in login. Currently we
use LDAP for authentication, and it populates a new user in database when
passes auth. That's normal. However if a user is existed in database and
"username" is different from LDAP user, there will be two different users
entries with the same email. So I need to check the user if exists by email,
and update username.